% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_regression.R
\name{backward_regression}
\alias{backward_regression}
\title{Backward Regression}
\usage{
backward_regression(dados, resposta)
}
\arguments{
\item{data}{base de dados que deseja fazer a regressão linear.}

\item{response}{variável resposta do banco de dados "data".}
}
\value{
Retorna o modelo de seleção Backward Regression, em que retorna todos os passos para a seleção do melhor modelo.
}
\description{
Irá pegar o banco de dados e a variável resposta que o usuário passar e irá aplicar o método Backward Regression e retornará o melhor modelo segundo esse critério de seleção.
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
