% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_grafico.R
\name{bic_grafico}
\alias{bic_grafico}
\title{BIC grafico}
\usage{
bic_grafico(dados, resposta)
}
\arguments{
\item{resposta}{variável resposta do banco de dados "data".}

\item{data}{base de dados que deseja fazer a regressão linear}
}
\value{
Retorna o gráfico de BIC com os melhores 5 modelos.
}
\description{
Irá pegar o banco de dados e a variável resposta que o usuário passar e irá retornar o gráfico BIC com os melhores 5 modelos.
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
