% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_Mallow.R
\name{Cp_Mallow}
\alias{Cp_Mallow}
\title{C_p de Mallow}
\usage{
Cp_Mallow(dados, resposta)
}
\arguments{
\item{resposta}{variável resposta do banco de dados "data".}

\item{data}{base de dados que deseja fazer a regressão linear.}
}
\value{
Retorna o gráfico de C_p de Mallow com os melhores 5 modelos.
}
\description{
Irá pegar o banco de dados e a variável resposta que o usuário passar e irá retornar o gráfico de C_p de Mallow com os melhores 5 modelos.
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
