% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Faz validação cruzada de modelos de regressão}
\usage{
cross_validate(modelos, dados, y, prop_train = 0.7)
}
\arguments{
\item{modelos}{Lista de modelos lineares}

\item{dados}{`Data.Frame` que está sendo usado na modelagem}

\item{y}{Nome da variável resposta}

\item{prop_train}{Proporção das observações de `dados` que será alocado para o grupo de treino, default é `0.7`}
}
\value{
Matriz com AIC, R^2, R^ Ajustado, MSE e MAPE de cada modelo
}
\description{
Faz validação cruzada de modelos de regressão
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
