% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_FRT.R
\docType{data}
\name{dadosFRT}
\alias{dadosFRT}
\title{Dados sobre Marketing'}
\format{
`dadosFRT`
Um data frame com 171 linhas e 4 colunas:
\describe{
  \item{youtube}{Vendas por Youtube}
  \item{facebook}{Vendas no Facebook}
  \item{newspaper}{Vendas no jornal}
  \item{sales}{Quantidade média de vendas.}
}
}
\source{
<https://www.kaggle.com/datasets/fayejavad/marketing-linear-multiple-regression>
}
\usage{
dadosFRT
}
\description{
Banco de dados que contém informações sobre marketing.
}
\keyword{datasets}
