% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_regression.R
\name{forward_regression}
\alias{forward_regression}
\title{Forward Regression}
\usage{
forward_regression(data, response)
}
\arguments{
\item{data}{base de dados que deseja fazer a regressão linear.}

\item{response}{variável resposta do banco de dados "data".}
}
\value{
Retorna o modelo de seleção Forward Regression, em que retorna todos os passos para a seleção do melhor modelo.
}
\description{
Irá pegar o banco de dados e a variável resposta que o usuário passar e irá aplicar o método Forward Regression e retornará o melhor modelo segundo esse critério de seleção.
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
