% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{modelos}
\alias{modelos}
\title{Cria combinações de modelos lineares}
\usage{
modelos(data, y, variaveis)
}
\arguments{
\item{data}{`data.frame` com variável resposta e covariáveis}

\item{y}{Nome da variável resposta do tipo `factor` ou `numeric`}

\item{variaveis}{Nomes das variáveis que podem entrar no modelo do tipo `factor` ou `numeric`}
}
\value{
Retorna uma lista de modelos lineares
}
\description{
Cria combinações de modelos lineares
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
