% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_descritiva.R
\name{tabela_descritiva}
\alias{tabela_descritiva}
\title{Tabela Análise Descritiva}
\usage{
tabela_descritiva(dados, variavel)
}
\arguments{
\item{variavel}{variável de interesse}

\item{data}{base de dados}
}
\value{
Retorna um data.frame formatado com métricas descritivas da variável de interesse.
}
\description{
Gera um data.frame com métricas descritivas para tabela.
}
\author{
Felipe Marques, Pedro Galera, Rubens Roncato, Tiago Monteiro e Vitor Rizzo.
}
